Adding some simple rendering test to get started
