+++ .gitignore
@@ -2,3 +2,4 @@
 /dist
 .env
 .idea
+/tests/chrome/
