+++ src/lib.rs
@@ -1,0 +1,4 @@
+
+pub mod pages;
+pub mod sl;
+pub mod route;
